requirements.txt
telethon==1.24.0
--openai==0.27.2
python-dotenv==1.0.0
asyncio==3.4.3

openai~=1.30.5
aiohttp~=3.9.5
PyYAML~=6.0.1